% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisePCoA.R
\name{pairwise_beta_plot}
\alias{pairwise_beta_plot}
\title{Pairwise PCoA plots}
\usage{
pairwise_beta_plot(s, dist_beta, color_by)
}
\arguments{
\item{s}{Metadata}

\item{dist_beta}{Distance object}

\item{color_by}{Grouping variable}
}
\value{
Pairwise PCoA plots
}
\description{
Pairwise PCoA plots
}
